Document ftp: have timeout at tests, skip if server is not available.
At trunk-xrg, the FTP daemon is only started *after* the db is loaded,
which means that it will not be available for tests on a server upgrade.
(still, tests can be performed through the base_module_record's importer)

Anyway, it is not wise to block the server startup if FTP doesn't
respond.