[FIX] issue with Firefox getting the whole template (including base JS/CSS) on some redirects (following button actions generally)
Issue comes from Firefox bug https://bugzilla.mozilla.org/show_bug.cgi?id=553888
When jQuery performs a request it sets the X-Requested-With header to XMLHttpRequest and we're using that for dispatching, so far so good.
But after a redirect (302, 303, whatever) while other browsers execute keep the X-Requested-With header around for the new GET request Firefox doesn't, and the server-side template dispatcher becomes confused, believing the request to be a 'full' one in need of everything even though it comes from XHR.
As a result, Firefox gets a bunch of files it re-reads and re-interprets binding a number of events and behaviors twice (or more) and creating a number of bugs as well as very slow loading for things which should be cheap and fast.

This was worked around by adding a custom 'requested_with' GET parameter during redirect if the X-Requested-With header is set, and upon receiving requests popping that parameter (if present) and setting X-Requested-With.