[FIX] breakage of the ajaxStop event dispatcher in case of /openerp/listgrid/get being called, therefore breakage of attrs and state changes management
Bug is actually in jQuery 1.4.2: complete() is called unconditionally during onreadystatechange when tiemout or readyState = 4 (request done). issue is, in case of JSONP requests the JSONP handler function *already* calls complete() (and success(), but success was already protected).

Result is that complete() is called twice after jsonp requests, which decrements jQuery.active (active requests counter) twice, and putting it into the negatives. Since jQuery.active is used to know when to fire ajaxStop (when it reaches 0), ajaxStop isn't fired anymore until the page is reloaded.

As with 1.4.3, switching to 1.4.4 seems impossible at this point (action buttons completely break for instance), so I backported the 1.4.3/1.4.4 fix instead (just wrap the complete() call in an 'if not jsonp conditional)