[FIX] document: search() must filter out documents in disallowed directories
Permissions on ir.attachments must not only respect the access rights
of the object they are attached to, but must also be careful about
directory permissions as introduced by the document module.

Minor fix in copy() method to avoid reading all fields of the
attachment including the binary data itself, when only the name
is needed. Also made the "(copy)" suffix translatable.