[FIX] incorrect rendering of hidden graphs
dhtmlx is apparently entirely unable to render invisible graphs
(hidden via display:none on itself or a parent, visibility: hidden
does not trouble it), it apparently uses the actual size of the graph
container in its rendering and display:none elements have no
dimensions.

A graph in an o2m in a secondary tab gets rendered in display:none
environment yielding a completely broken and unusable graph.

Worked around this by deferring the graph's rendering until
visibility: the actual graph rendering is stored in a function which
is periodically deferred (via setTimeout) until we know the graph view
is visible (meaning its root element isn't invisible
anymore). Ideally, we'd use a library which is not broken instead...

The check periodicity is currently set to 100ms, seems sufficiently
reactive for our purposes.