[FIX] avoid having the XML concatenator blow up when passed an empty list of files                                                                                     ~/projects/tiny/web/current
turns out ElementTree has a hard time XML-serializing ``None`` (who'd
have guessed...), which is what it gets told to do when concat_xml is
called with an empty list of files.

Return an empty string and no timestamp (should probably be one there
at some point, but...) and fix the QWeb loader to not do anything when
it gets an empty template file.