[FIX] fetching of export data after reloading from saved list in a different mode than original
If a field name can not be found in the `records` mapping, use it
as-is. May fail in some cases (for nested id fields) but will
generally work better when reloading a list created in "all" mode
while in "export compatible" mode, as there may be saved fields
missing from the currently available list.

Better fixes would be:

* Ignore (gray out?) unavailable fields for this mode (and re-allow
  when the mode changes)
* Display an error message of some sort and refuse to load the list
* Pre-load all export lists and filter selectable ones by mode