UF-1184 [FIX] Obviously few details need to be modified/added:
On Items lines: I prefer to be consistent with system's screens:
Items should be renamed Lines and the sequencing number also to avoid confusion.
A supplier code field should be added if we have any
The general note of the PO should be appearing on remarks field if any.
The remarks shouldn't have filled like this. Indeed it was indicated that the instance can optionally required to get the sentence "All order are done...on request" (because it is not always the case).
Please also remove the "please identify all parcels..." that was for comment to the dev (but we'll manage reference sequence convention in a different UF).