UTP-714 [FIX]
2. Please add the PO reference on the "import PO wizard"
4. The note on "Import po" after the load is wrong and should be consistent with updated information
6. Please modify the message while the system is uploading the PO content by:
" Please note that POXXX is temporary closed during the import to avoid conflict accesses (you can see the loading on the PO note tab check box). At the end of the load, POXX will be back in the right state. You can refresh the screen if you need to follow the upload progress"