UTP-760 [IMP] Issues with HQ entries Split (analytical and accounting)
Since UTP-727 when a wizard is closed from a tree view, the selected ids are kept checked.
If the wizard deletes some ids, it crashes.
To solve this issue, when a wizard is closed and if ids was previously selected, we check if
previous ids still exist and match the current view domain.