UF-2236 [FIX] OCB export - problem with rawdata result that show lines from other periods
When you export OCB lines, you choose a period. This one was not used for "rawdata" SQL request because of analytic lines that have no period.

The tip is to take first day of period and last day from period and use the "date" field from analytic lines to select only lines from a given period.