UF-2281 [FIX] Problem with double creation of move line from a register line
We create a move line from st line if line is temp posted (a normal line). Which is wrong.

We should so return to a previous state of the code for "normal" lines and keep the new behaviour for direct invoice.

More explanation about use cases is available in the code at posting() method from account_bank_statement_line class.