UF-1746 [FIX] Bug on analytic lines that are corrected in closed period: No REV and COR lines.
This is due to the UF-1746 branch in which the "def corrected_upstream" make a write on account move line which delete and recreate analytic lines.

As this write permits to indicate synchronization that a change was done, I use "update_check=False, check=False" to avoid analytic lines deletion/recreation. Which solves the problem.