[FIX] Problem with some analytic lines that are not deleted during synchronization.
It appears that in synchronization, @update, we use a dictionary to find some values. In a dict, values are not sorted, so if we make a couple of value with keys from 2 dictionary, they are not link to the right ID/key/value.
It's now fixed thanks to Jean-François.