UF-2377 [FIX] Bad USB synchronization of packing list on shipment
At USB synchronization of a shipment, if two packing list have the same PPL name
sometimes, at Central Platform, the product of the two packing lists were       
switched. There was a problem because we didn't check if we put the selected
number value to the good packing list.
Now, we check if the couple from pack/to pack is the same as the PPL to put in
the shipment.