UFTP-185 [FIX]
In the Partner selection box we can see Suppliers + Customers.
However, as this report is the FO follow-up per Client, only Customers should be seen.

Order ref : might it be possible to filter them to show only the ones from the Partner that has been selected, if one has been selected ? Indeed, after some months of use, we will have a very big number of orders and this filter will be very useful.