UFTP-185 [FIX]
Regarding the over-shipment (negative BO) I agree that we need to show the fact that we sent more, but I still do think that we cannot show negative BOs. Maybe we could use the following : Put 0 as BO (because there are no BO and no additional qty will be shipped) and in brackets the additional qty shipped : BO : 0 (+25).