US-713 [ADD] new features. This is an intermediate commit to save my work  before the WE. This is not finished work. More commits to comme.
[ADD] a new wizard wich permit to create temp object corresponding to the
invoice and invoice_lines. It is yet possible to add, delete, modify this
lines from the wizard, but not possible to save/validate. If the wizard is
closed without saving, nothing is saved (this is exacly the behaviour we are
looking for)

[ADD] two classes to represent the temp invoice and temp invoice_lines

For now a lot of code is duplicated and some functionnalities are not
implemented (Analytic Distributions button, Compute Total button and Validate
button for example).