SP-223 [IMP] declare execution_date to avoid to have the case of some update with a lower id have an execution_date after an higher id. Here is an example :
logs.keys() = [1,2,3,4]
update_ids = [1,2,3,4,5,6,7]

ids_not_in_logs = [5,6,7]
In this case, we will first write ids_not_in_logs [5,6,7] with current execution_date
and after that we will write logs.keys() [1,2,3,4]. This mean that the update
with id=1 will have an execution_date after the update with id=5.

This is not really a problem for normal work, but for performance measurement
this give negative execution time.