SP-222 [FIX] do not group updates if the fields are not the same
This could be the case if the fields have been modified or some fields
deleted/added int the rule. In this case it is not possible to group them
because then the sync client will receive updates with wrong fields/values.
This case was happening for the account.period.state where there were some
updates with "period_id" fields and some other with "period_id/id". This
should not be grouped.