SP-222 [FIX] do not move update to lower id if they refere to object modified/created in an higher update id (ie. in the "future")
For example
update 1:
creation of a partner PAR
update 2:
creation of a product PRO refering to partner PAR
update 12:
creation of partner PAR2
update 13:
modification product PRO to refere to partner PAR2

In this case, it is not possible to compact by replacing content of update 1
by content of update 13 because then the partner PAR2 will not exists at this
time.


[IMP] change the execution order of the different steps. Searching if the is
related object in the Future before to move update is a costly operation. It is
much more efficient to first remove all the update "easy to remove" such as
- the ones with active rules and not master_data
- the ones related to inactive rules
then there is much less update to parse with costly operation (about 6,8
Million less !)

[IMP] add constant to be more configurable