US-1207 [IMP] call a specific method approximate_search_count() instead of the original search_count(). The count result returned is now always an int. A boolean is also returned in addition to know if the result is approximated or not.
In case the result is approximated, do not display the pager arrow to go to
the last page of the result. Because the number of objects is approximated, the
last_page is hazardous to get.
Over more 100000 result, there is no sens to go to the last page, result can be
reordered differently (ordered by create_date desc for example) and filter can
be added to restrict the search scope to a smaller set.