US-1267 [FIX] Some procurement order moves were done and with 0.00 as product_qty and this blocked the processing of an Incoming Shipment if one of these move was for a perishable/batch mandatory product.
Description of the issue:
When lines are added to a PO by the sync. engine (due to lines added on the FO at other sync. side), these lines generates a procurement order and a stock move with Procurement as source location.
But when the sync. engine changes the state of the Incoming shipment from 'Available' to 'Available Shipped', the stock move associated to the procurement order is put as 'Done' and without quantity.
With this fix, the procurement order will get the associated stock move of the IN.