US-1191/2 [REF]
UC: if you remove the periods after the fiscal year is selected,
the "include initial balances" button becomes unticked and grey.

The box should be able to be selected when periods are left blank
(=all periods to date are included)
or when the "From" period is set to January.
This is the behaviour in G/L and Trial Balance. 

=> refactor wizard and allow to input empty period from/to frontiers
=> only disable IB tickbox if month <> Jan (allow empty frontier)