US-1246 [IMP] _get_batch_location_ids method was very slow. Improve the code to get it fast. The above call : ids = stock_move.read([71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100], ['line_number', 'currency', 'uom_id', 'cost', 'kc_check', 'ordered_quantity', 'ordered_uom_id', 'composition_list_id', 'asset_id', 'ssl_check', 'type_check', 'location_id', 'np_check', 'ordered_uom_category', 'integrity_status', 'exp_check', 'prodlot_id', 'expiry_date', 'product_id', 'lot_check', 'batch_location_ids', 'dg_check', 'kit_check', 'asset_check', 'quantity', '__last_update'], {'lang': u'en_MF', 'client': 'web', 'tz': False, 'department_id': False})
was generating 1151 requests in _get_batch_location_ids before this patch, only 8 after the patch.

[IMP] to do this improvments, I also introduced the name_get parameter in some methods (especially read()).
name_get calculation is sometimes costly and most of the times we don't need it. By default it is set to True to keep old behaviour, but if it is set to False, name_get will not been called and a tuple (id, False) will be returned instead.
[IMP] small minor changes