US-1246 [IMP] remove total_amount field that was used only in one report, and this report was never used. Remove this report too after checking with Quentin. Improve the way stok.picking._vals_get() method works to be faster. Now it returns only the values asked, not all values. This fields are not anymore mutli.
same work remains on stock.move._vals_get
it seems that a log of report (and so a lot of code, method and properties) are
useless. See bin/addons/msf_outgoing/msf_outgoing_report.xml for commented old
reports.

This properties sometimes relates on fields.function which are very slow and
can be read outside of this useless reports. So it can slow down other parts of
unifield.