US-1246 [IMP] significant improvment of _shipped_rate method. Before it was doing thousands or sql request each time the listbox of Purchases were displayed (to display the progress bar), now the same job is done bug with only 4 reads.
I added the method _compute_qty_dict and _compute_qty_obj_dict that can handle
a dict as parameter, this avoid to look for information we already know and
save a lot of requests/time machine.

I added a XXX comment because I don't understand the purpose of some code (that
I deleted in the dict version.

A futur improvment could be to change all other call to _compute_qty and then
keep only the dict version.