US-1246 [IMP] this is a test to check if everything is working by reversing the list :
instead of doing a MODELS_TO_IGNORE list, I did the oposite, get all models we
need for sync (extracted from sync_server_sync_rule and
sync_server_message_rule) and remove them from the list of all modules. The
result is now the new MODELS_TO_IGNORE list.

Everything is hardcoded here but should be not in the future.