US-1381 [IMP] encrypt all password This is transparent for the user, but now all new user will have an encrypted password, and all existing users that will change their password (via force_password_change or manually) will have an encrypted password.
The old password are still working : the framework is able to detect if the
password is clear or encrypted and do the adequat check. But it is not possible
to define a new unencrypted password via the web interface (only with sql
commands).

About the technology used it is SHA-512 with randomize salt. This is quite
safe without the pain of adding more library to unifield.