[FIX] US-3149: fix failure formatting report
reportlab 3.4.0 wants to eat all the insides of a <para>
at the same time. With the old code, feeding it
<para><font color="white">\n</font></para> would fail because
the <font> would appear to not be closed.

In engagement.rml, we have <font color="white"/>, which is
valid. However, at another place in the code, the
RML is regenerated, pretty-printing the <font/> into
<font color="white">\n</font>, and causing the problem.