[FIX] US-3149: Set padding correctly after reportlab update
With the new reportlab, reports were being formatted incorrectly.
The root problem seems to be that the default padding for left/right
used to be 0 and changed to 6. This change makes all users of
RML get a default of 0 left/right again.

Also: Updated the import for Image, according to the docs for
Pillow.