US-3085 [ADD] when automatic sync is modified,ensure that the automatic sync is in the time range of the silent upgrade if defined. And when the silent upgrade is modified, check its time range match the time of the automatic sync.
In case it don't match check for the next nextcall during the next 7 days if
any match. I none match during this 7 days, raise an error.

This done like that to be able to work in the following case:
Automatic sync defined at 08h00 AM every 12 hours
Time range of the silent upgrade : from 19:00 to 21:00

Whith the attached patch, this example will work.