[FIX] US-3506: Remove a dependency that may exist and stop upgrade
OCA has a hand-added dependency of OpenERP Server on Postgres.
This causes the upgrade to fail because "net stop Postgres_for_OpenERP"
asks for confirmation that OpenERP Server should be stopped as well,
and as this is in batch mode, there is no human to type "y".

The solution is to remove the dependency, since that is how OpenERP
Server has always run on all the other instances, and because
we know that OpenERP Server (via it's ORM) is capable of reconnecting
until it finds that Postgres is up.