US-8094 [IMP] HQ Entries: add Track Changes
Notes on some confusing fields:
user_validated: renamed to "Validated". This field isn't displayed.
is_original: renamed to "Has been split". This field is used in the interface but its name is overwritten.
is_split: doesn't refer to a line being split but to a line generated by a "split". Not tracked. The field "Original HQ Entry", which is filled in for those lines, will be used instead.